requirements.txt
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0
pyyaml>=6.0
dpkt>=1.9.8
scapy>=2.4.5
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=3.0.0
structlog>=22.1.0
requests>=2.28.0
python-dateutil>=2.8.2
